+++ backend/routes/routes/schoolsRoutes.js
@@ -6,23 +6,27 @@ const checkAuth = require('../../utils/check-auth');
 // you need to set mergeParams: true on the router,
 // if you want to access params from the parent router
 var router = express.Router({mergeParams: true});
-var defineRoutes = router =>{
+var defineRoutes = router => {
 
-    router.get('/:id/cms', async (req,res)=> {
+    router.get('/:id/cms', async (req, res) => {
         let result = await SchoolController.cms(req.params.id);
-        res.status(result?200:400).send(result);
+        res.status(result ? 200 : 400).send(result);
     });
 
-    router.get('/:id', async function(req,res){
-        let result =  await SchoolController.getSchool(req.params.id);
-        res.status(result?200:400).send(result);
-    });
-
-    router.post('', checkAuth, async function(req,res){
+    router.get('/:id', async function (req, res) {
         let result = {};
-        if(!req.body.institutionid) {
+        if (!req.params.id)
             result = {error: true, description: 'you don\'t have validation'};
-        }else {
+        else
+            result = await SchoolController.getSchool(req.params.id);
+        res.status(result ? 200 : 400).send(result);
+    });
+
+    router.post('', checkAuth, async function (req, res) {
+        let result = {};
+        if (!req.body.institutionid || !req.body.name) {
+            result = {error: true, description: 'you don\'t have validation'};
+        } else {
             var send = {};
             send.institutionid = req.body.institutionid;
             result = await SchoolController.createSchool(req.body);
@@ -31,44 +35,69 @@ var defineRoutes = router =>{
                 InstitutionController.addSchool(send);
             }
         }
-        res.status(result?201:400).send(result);
+        res.status(result ? 201 : 400).send(result);
     });
 
-    router.get('', async function(req,res){
-        let result =  await SchoolController.getSchoolCollection(req.body);
-        res.status(result?200:400).send(result);
+    router.get('', async function (req, res) {
+        let result = await SchoolController.getSchoolCollection();
+        res.status(result ? 200 : 400).send(result);
     });
 
-    router.delete('', checkAuth, async function(req,res){
-        let result =  await SchoolController.deleteSchool(req.body._id);
-        if(!result.error)
-            InstitutionController.deleteSchool({institutionid:req.body.institutionid,schoolid:req.body._id});
+    router.delete('', checkAuth, async function (req, res) {
+        let result = {};
+        if (!req.body._id || !req.body.institutionid)
+            result = {error: true, description: 'you don\'t have validation'};
+        else {
+            result = await SchoolController.deleteSchool(req.body._id);
+            if (!result.error)
+                InstitutionController.deleteSchool({institutionid: req.body.institutionid, schoolid: req.body._id});
+        }
         res.status(200).send(result);
     });
 
-    router.put('', checkAuth, async function(req,res){
-        let result =  await SchoolController.updateSchool(req.body);
+    router.put('', checkAuth, async function (req, res) {
+        let result = {};
+        if(!req.body._id)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await SchoolController.updateSchool(req.body);
         res.status(200).send(result);
     });
 
-    router.post('/addsubject', checkAuth, async function(req,res){
-        let result =  await SchoolController.addSubject(req.body);
-        res.status(result?201:400).send(result);
+    router.post('/addsubject', checkAuth, async function (req, res) {
+        let result = {};
+        if(!req.body.schoolid || !req.body.subjectid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await SchoolController.addSubject(req.body);
+        res.status(result ? 201 : 400).send(result);
     });
 
-    router.get('/getsubjects/:id', async function(req,res){
-        let result = await SchoolController.getSubjects(req.params.id);
-        res.status(result?200:400).send(result);
+    router.get('/getsubjects/:id', async function (req, res) {
+        let result = {};
+        if (!req.params.id)
+            result = {error: true, description: 'you don\'t have validation'};
+        else
+            result = await SchoolController.getSubjects(req.params.id);
+        res.status(result ? 200 : 400).send(result);
     });
 
-    router.post('/addpermission', checkAuth, async function(req,res){
-        let result = await SchoolController.addpermission(req.body);
-        res.status(result?200:400).send(result);
+    router.post('/addpermission', checkAuth, async function (req, res) {
+        let result = {};
+        if(!req.body.schoolid || !req.body.userid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await SchoolController.addpermission(req.body);
+        res.status(result ? 200 : 400).send(result);
     });
 
-    router.delete('/permission', checkAuth, async function(req,res){
-        let result = await SchoolController.deletepermission(req.body);
-        res.status(result?200:400).send(result);
+    router.delete('/permission', checkAuth, async function (req, res) {
+        let result = {};
+        if(!req.body.schoolid || !req.body.userid)
+            result = {error:true,description:'you don\'t have validation'};
+        else
+            result = await SchoolController.deletepermission(req.body);
+        res.status(result ? 200 : 400).send(result);
     });
 
     return router;
